requirements.txt
streamlit
streamlit-chat
langchain
langchain-community
langchain-groq
langchain-google-genai
chromadb
python-dotenv
pymongo
streamlit-option-menu
sentence-transformers
email-validator
pandas
qdrant_client
langchain_chroma
pymongo
langchain_huggingface
textblob
google-generativeai
langchain_qdrant
qdrant_client